add auth command and actions

+++ app/Http/Controllers/ChirpController.php
@@ -4,8 +4,10 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Chirp;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class ChirpController extends Controller {
+    use AuthorizesRequests;
     /**
      * Display a listing of the resource.
      */
@@ -41,9 +43,8 @@ class ChirpController extends Controller {
         'message.max' => 'Chirps must be 255 characters or less.',
     ]);
 
-        Chirp::create([
-            'message' => $validated['message'],
-        ]);
+        auth()->user()->chirps()->create($validated);
+
 
         return redirect('/')->with('success', 'Chirp created!');
     }
@@ -61,6 +62,7 @@ class ChirpController extends Controller {
      */
     public function edit(Chirp $chirp)
     {
+         $this->authorize('update', $chirp);
         return view('chirps.edit', compact('chirp'));
     }
 
@@ -69,6 +71,8 @@ class ChirpController extends Controller {
      */
     public function update(Request $request, Chirp $chirp)
     {
+
+        $this->authorize('update', $chirp);
         $validated = $request->validate([
             'message' => 'required|string|max:255|min:5',
         ],
@@ -87,6 +91,7 @@ class ChirpController extends Controller {
      */
     public function destroy(Chirp $chirp)
     {
+         $this->authorize('update', $chirp);
         $chirp->delete();
         return redirect('/')->with('success', 'Chirp deleted!');
     }
